combine daily and days per week
